Extract FileMenu.java
